requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
